fix dial tests
One *can* write to a closed connection. It'll just fail eventually (but not
unless you send a bunch of data). The better test is to read from it.

Note: We only accept connections once they've been secured. This test *used* to
pass because we'd race and close the connection *before* the handshake
completed.