Fix a goroutine leak in accepting connections
If a connection fully establishes, passing the timeout select in
handleIncoming, but then is never taken out of a full incoming queue,
the goroutine is leaked. handleIncoming will in turn leak waiting for
wg.Done. Canceling the context, closing the listener, or closing the
goprocess don't affect these goroutines. connAcceptBuffer makes this
problem a little less evident, but doesn't fix it.